aqm-backend/requirements.txt
pandas
lightgbm
scikit-learn
joblib
numpy
datetime
flask
psycopg2-binary
